+++ src/ui/EditServersPopup.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "Types.hpp"
+
+class EditServersPopup : public geode::Popup<std::vector<GDPSTypes::Server>> {
+    protected:
+    bool setup(std::vector<GDPSTypes::Server> servers) override;
+
+public:
+    static EditServersPopup* create(std::vector<GDPSTypes::Server> servers);
+};