+++ src/ui/ServerListLayer.hpp
@@ -13,7 +13,6 @@ class ServerListLayer : public cocos2d::CCLayer {
         bool init() override;
         void keyBackClicked() override;
 
-        static GDPSTypes::Server m_selectedServer;
 
         cocos2d::CCMenu *m_backMenu = nullptr;
         geode::ScrollLayer *m_scroll = nullptr;
@@ -24,9 +23,14 @@ class ServerListLayer : public cocos2d::CCLayer {
     public:
         static ServerListLayer *create();
         static cocos2d::CCScene *scene();
+
+        void updateList();
         
         void onBack(CCObject *sender);
+        void onAdd(CCObject *sender);
+        void onEdit(CCObject *sender);
         void onSelect(GDPSTypes::Server server);
 
-        GDPSTypes::Server getSelected();
+        static GDPSTypes::Server m_selectedServer;
+        std::vector<GDPSTypes::Server> m_servers;
 };